refactor: constants 파일 네이밍 수정
