+++ src/constants/housing-options.ts
