+++ src/constants/listing-card-dummies.ts
