+++ src/constants/listing-detail-tabs.ts
